Move toolbar to own component
